+++ README.md
@@ -1,0 +1,8 @@
+## Shiken
+
+Is a collection of packages that help with setting up testing harnesses.  They are designed with the user in a monorepo in mind, but it is not limited to that.  
+
+Uses [Flutter's Widget tests](https://docs.flutter.dev/cookbook/testing/widget/introduction) under the hood.
+
+It uses [Given, When, Then](https://en.wikipedia.org/wiki/Given-When-Then) to describe tests.  
+
