+++ packages/shiken_test_harness/README.md
@@ -1,0 +1,28 @@
+# Test Harness
+
+[![style: very good analysis][very_good_analysis_badge]][very_good_analysis_link]
+[![Powered by Mason](https://img.shields.io/endpoint?url=https%3A%2F%2Ftinyurl.com%2Fmason-badge)](https://github.com/felangel/mason)
+[![License: MIT][license_badge]][license_link]
+
+Helps setup a test harness
+
+## Installation 💻
+
+**❗ In order to start using Test Harness you must have the [Flutter SDK][flutter_install_link] installed on your machine.**
+
+Add `test_harness` to your `pubspec.yaml`:
+
+```yaml
+dev_dependencies:
+  test_harness:
+```
+
+Install it:
+
+```sh
+flutter packages get
+```
+
+## Usage 
+
+-- todo
