+++ packages/shiken_test_harness/lib/src/harness_setup/unit_test_harness_setup.dart
@@ -1,4 +1,4 @@
-import 'package:test_harness/test_harness.dart';
+import 'package:shiken_test_harness/test_harness.dart';
 
 /// Used to create a Unit test harness
 ///
