+++ packages/shiken_test_harness/lib/src/harness_setup/widget_test_harness_setup.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:test_harness/test_harness.dart';
+import 'package:shiken_test_harness/test_harness.dart';
 
 /// Faciliates the creation of a [WidgetTestHarness] function for use in a widgetTest
 ///
