+++ packages/shiken_test_harness/test/src/example/example_unit_harness.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:test_harness/test_harness.dart';
+import 'package:shiken_test_harness/test_harness.dart';
 
 import 'counter_harness_mixin.dart';
