+++ packages/shiken_test_harness/test/src/example/harness_mixins/counter_harness_mixin.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
-import 'package:test_harness/test_harness.dart';
+import 'package:shiken_test_harness/test_harness.dart';
 
-import 'example_model.dart';
+import '../example_model.dart';
+
 
 mixin CounterHarnessMixin on TestHarness {
   CounterModel counter = CounterModel();
