+++ packages/shiken_test_harness/test/src/test_harness_test.dart
@@ -3,7 +3,7 @@
 import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:test_harness/test_harness.dart';
+import 'package:shiken_test_harness/test_harness.dart';
 
 void main() {
   group('TestHarness', () {
